Use correct builders for `ServerEventStreamMarshallerGeneratorTest`
